split asset submodule
